+++ src/main/java/org/cowboycoders/ant/profiles/simulators/TurboStateViewable.java
@@ -32,5 +32,5 @@ public interface TurboStateViewable {
 
     Integer getHeartRate();
 
-    int getDistance();
+    long getDistance();
 }
